requirements.txt
django
djangorestframework
pandas
openpyxl
gunicorn
django-cors-headers
